requirements.txt
Pillow==8.0
googlevoice==2.0.5
